requirements.txt
python-dotenv
pandas
numpy
pymongo
scikit-learn
certifi
pymongo[srv]
-e .
